Clean Code Refactored Payload to the client package
