+++ mpesa-mobile-payment-impl/src/main/java/com/kwawingu/payments/c2b/CustomerToBusinessTransaction.java
@@ -6,9 +6,9 @@ package com.kwawingu.payments.c2b;
 import com.kwawingu.payments.ApiEndpoint;
 import com.kwawingu.payments.Service;
 import com.kwawingu.payments.client.MpesaHttpClient;
+import com.kwawingu.payments.client.payload.Payload;
 import com.kwawingu.payments.session.keys.MpesaEncryptedSessionKey;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.util.HashMap;
